requirements.txt
python-novaclient
python-neutronclient